refactor: split Drive client into auth, gateway, and source layers
DriveClient was a God Object owning four unrelated responsibilities:
auth/credential lifecycle, service transport and retry, Drive source
semantics, and payload parsing.

Split into three explicit layers with a narrow protocol:
- drive_types.py: shared dataclasses and errors (DriveFile, DriveAuthError, etc.)
- drive_auth.py: DriveAuthManager + DriveAuthPrompter protocol + path helpers
- drive_gateway.py: DriveServiceGateway (import, service lifecycle, retry, download)
- drive_source.py: DriveSourceClient + DriveSourceAPI protocol + pure helpers

drive_client.py becomes a thin composition root delegating to these layers.
drive.py now depends on DriveSourceAPI (protocol) instead of the concrete class.
auth.py now calls DriveAuthManager.load_credentials() directly — the
resolve_folder_id("root") auth hack is removed.

Test ownership reassigned:
- test_drive_auth.py: credential loading state machine, path resolution, prompter
- test_drive_gateway.py: retry policy, service caching, download transport
- test_drive_source_client.py: folder resolution, iteration, payload parsing, download semantics
- test_auth.py: slimmed to CLI routing and DriveAuthManager dispatch

Deleted: test_drive_client_laws.py, test_drive_utils.py (fully absorbed by owners).
All 2689 unit tests pass.
